requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
realesrgan==0.3.0
